+++ floes/NonEquilSwch.py
@@ -47,23 +47,24 @@ coll_open.set_parameters(write_new_collection='NES_OPLMD')
 coll_close = CollectionSetting("CloseCollection", title="Close Collection")
 coll_close.set_parameters(open=False)
 
-rec_check = ParallelRecordSizeCheck("Record Check Success")
+rec_check = ParallelRecordSizeCheck("Record Check Success", title="Record Size Checking")
+rec_check_abfe = ParallelRecordSizeCheck("Record Check Success ABFE", title="Affinity Record Size Checking")
 
 ofs_nes = DatasetWriterCube('ofs', title='NES Out')
 ofs_nes.promote_parameter("data_out", promoted_name="out",
                           title="NES Dataset Out",
                           description="NES Dataset Out", order=4)
 
-ofs_abfe = DatasetWriterCube('ofs_abfe', title='ABFE Out')
+ofs_abfe = DatasetWriterCube('ofs_abfe', title='Affinity Out')
 ofs_abfe.promote_parameter("data_out", promoted_name="abfe",
-                           title="ABFEt Out",
-                           description="Absolute Binding Affinity Out", order=5)
+                           title="Affinity Out",
+                           description="Affinity Out", order=5)
 
 fail = DatasetWriterCube('fail', title='NES Failures')
 fail.promote_parameter("data_out", promoted_name="fail", title="NES Failures",
                        description="NES Dataset Failures out", order=6)
 
-job.add_cubes(iun, ibn, coll_open, coll_close, rec_check, ofs_nes, ofs_abfe, fail)
+job.add_cubes(iun, ibn, coll_open, coll_close, rec_check, rec_check_abfe, ofs_nes, ofs_abfe, fail)
 
 nes_subfloe_options = dict()
 nes_subfloe_options['edge_map_file'] = 'map'
@@ -73,7 +74,7 @@ nes_subfloe_options['nes_switch_time_in_ns'] = 0.05
 input_port_dic = {'input_open_collection_port': coll_open.success,
                   'input_bound_port': ibn.success}
 output_port_dic = {'output_nes_port': coll_close.intake,
-                   'output_abfe_port': ofs_abfe.intake,
+                   'output_abfe_port': rec_check_abfe.intake,
                    'output_fail_port': rec_check.fail_in}
 
 nes_gmx_subfloe(job, input_port_dic, output_port_dic, nes_subfloe_options)
@@ -82,10 +83,12 @@ iun.success.connect(coll_open.intake)
 ibn.success.connect(coll_open.intake)
 coll_close.success.connect(rec_check.intake)
 rec_check.success.connect(ofs_nes.intake)
+rec_check_abfe.success.connect(ofs_abfe.intake)
 
 coll_open.failure.connect(rec_check.fail_in)
 coll_close.failure.connect(rec_check.fail_in)
 rec_check.failure.connect(fail.intake)
+rec_check_abfe.failure.connect(fail.intake)
 
 if __name__ == "__main__":
     job.run()